requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
python-jose
passlib
sqlalchemy
python-jose[cryptography]
python-dotenv
bcrypt
httpx
typing